AND-6195 Move AddHeaderInterceptor to new package
