+++ blockchain/src/main/java/com/tangem/blockchain/blockchains/hedera/network/HederaMirrorRestProvider.kt
@@ -2,8 +2,8 @@ package com.tangem.blockchain.blockchains.hedera.network
 
 import com.tangem.blockchain.common.Blockchain
 import com.tangem.blockchain.common.BlockchainSdkError
+import com.tangem.blockchain.common.logging.AddHeaderInterceptor
 import com.tangem.blockchain.common.toBlockchainSdkError
-import com.tangem.blockchain.extensions.AddHeaderInterceptor
 import com.tangem.blockchain.extensions.Result
 import com.tangem.blockchain.extensions.retryIO
 import com.tangem.blockchain.network.createRetrofitInstance
